split `sunburst` into submodules `chem_env`, `chem_sys`, and `spacegroup`
